ui: extract TrackEvent callstack flamegraph tab
Move the TrackEvent callstack flamegraph area-selection implementation into a dedicated tab class without changing its behavior.
